delete movie list added.

+++ classes.py
@@ -302,9 +302,9 @@ class MovieList:
     def delete_list(self):
          with dbapi2.connect(app.config['dsn']) as connection:
                     cursor = connection.cursor()
-                    query = """DELETE FROM MOVIELIST WHERE (LIST_NAME = %s)"""
+                    query = """DELETE FROM MOVIELIST WHERE ((LIST_NAME = %s) AND (USER_ID = %s))"""
 
-                    cursor.execute(query, (self.list_name,))
+                    cursor.execute(query, (self.list_name, self.user_id,))
                     connection.commit()
 
     def exists(self):
